[Dronecourse] Add PlatformLanding message and ability to move platform
This commit adds the ability to move the platform using the platform
utility tool.

This commit also adds a contact sensor to the platform. The drone
velocity is monitored and published with a flag which is set if the
drone is in contact with the platform. This PlatformLanding message
is then sent over MAVLink protocol. This message allows students to
determine the drone's velocity upon landing on the platform.